refactor(dpp): split document serialization formats into per-generation files
serialize.rs had grown to 3,815 lines holding four frozen wire-format
generations plus their tests. Each format now lives in its own file —
document/v0/serialize/{v0,v1,v2,v3}.rs — as inherent methods with bodies
moved verbatim; the trait impls in mod.rs are one-line dispatch shims, and
the tests move unchanged (one dedent) into tests.rs. Pure relocation: git
diff --color-moved-ws=allow-indentation-change classifies all but 85 of the
~4,000 added lines as moves, the 85 being module plumbing (headers, impl
wrappers, pub(super) markers, shims, test imports).

The point is consensus hygiene: a shipped format's file must never change,
so a future diff touching serialize/v1.rs is immediately suspect in review
— the same physical isolation the version tables give generation modules.
